Update library components to add description
Added descriptions for the LS7366R and VNH5019ATR-E.
Also includes correct pin count information (apparently it needs the Pin
Count parameter set, not the Number of Pins parameter from Altium
Content Library, hah).  Added supplier and pricing info to the LS7366R
(not that it helps in the BOM really, or the SCH Library supplier links
list).