General check-in
Been messing around a lot, want to commit the current working version.
I've renamed the TOC but I think I want to rename it again; sadly the
Storage Manager sees a pending change and won't let me rename it within
the project, so I need to commit this before renaming it again.