Complete driver-counter schematics
This should finish off the driver-counter schematic sheet.  I went with
a 40MHz crystal because why not.  The load capacitance of that Fox
crystal is 12pF, which brings the capacitors to a 12pF-14pF range (see
LS7366R datasheet) depending on whether there's any extra coupling
capacitance between the crystal and ground.