Add an assortment of Molex SL Modular connectors
As it turns out, I didn't actually need to model the 70543 (no peg)
connectors because those will be added to the Pi HAT and aren't actually
on my PCB, oops.  Well, maybe they'll be handy later on.  Most of the
work was in implementing the 8-pin 70545-0077 anyway, the rest was just
copy-pasting and ripping pins and pegs out.