Update schematic templates, add misc. connections
This update adds the Miscellaneous Connectors schematic sheet where
hopefully I can break out the SPI and I2C buses for expansions as well
as some GPIO pins from the Mega.

The LED Driver and SBC Interface sheets are currently incomplete but
should be pretty quick - the LED Driver is just a header for the
Adafruit PCA9685 breakout PCB and the SBC Interface is just connectors
and a TXB0108.

Finally, I actually went through and ensured all of the schematic
sheets, old and new, are using my template and have their page
properties set correctly.