Assign real devices to (almost) all components
With the exception, I think, of the sense resistors, all components are
selected.  The INA199s and the RT9069 don't have models yet, and the IPC
Footprint Wizard doesn't get the idea of a 5-pin SOT-89, so it's
probably a good time to switch to a LM317 anyway!

Trying to import what I've got to the PCB seems like it works very well,
I just need to implement those remaining models... but for some reason
the TXB0108 (which is properly implemented as far as I can tell) also
throws an error.  Maybe that'll resolve itself once I tie off absolutely
every last loose end, we'll see.

Also I switched the so-so green power LED out for a very bright blue GaN
LED, because why not.