Tweak layer stackup, re-export rules with AD 19.1
I've been investigating impedance profiles and fine-tuning the layer
stackup manager, trying to get a feel for tolerances and how to make
100Ω traces.

Also, it looks like rule import/export is fixed in Altium 19.1, so I
re-exported the rule set on master and was able to re-import it without
too much trouble.  Hooray!  I can now set up separate rulesets for both
Seeed (0.1mm and 0.4mm) and Elecrow and anyone else.